+++ CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Issue [#546](https://github.com/m-marini/wheellyj/issues/546): Mapping state
+
 ### Changed
 
 - Issue [#545](https://github.com/m-marini/wheellyj/issues/545): Point echo
-
 
 ## [0.23.0] 2025-06-26